Fix problem when clicking ``Register for Classes'' where an XHTML validation error was thrown.
This error was caused by the ajax request using the HTTP GET method instead of POST. The HTTP GET method has a limitted length and the serialized querystring was much longer than that limit, causing fun errors. Perhaps this querystring limit was even on the browser side instead of apache-side... or perhaps apache was returning an error page to the browser about the querystring.