Add basic crawler support for Hope College in light of their winning the Calvin v. Hope basketball game last weekend.
The crawler is unable to, from the data provided by hope, properly recognize
and handle course labs in any fashion. To do that, the human-formatted comment
fields will need to be parsed. These fields are placed in the same column as
the course title is and are in the subsequent row for a particular section --
these fields are currently ignored through the rule which throws out rows with
too few columns.