Don't disable autocomplete for schools which failed to crawl in the last rehash.
This is especially important for hope as http://plus.hope.edu/ seems to always 503
in the middle of the night when we crawl it. Before this commit, that means that the
otherwise valid crawl data associated with hope would be available in cache/ but
would be ignored because hope was marked as uncrawled.