hope: Fix crawling of ACCT (Accounting) subjects for Hope College (thanks to Cody Stuive for reporting this!)
Apparently the CSV endpoint of the Oracle PROD thing actually *does*
require the "dummy" value to be sent in as the first sel_subj form
value. The CSV thing just ignores the first sel_subj even though the
HTML output version does not (and I assumed that the CSV form in the
HTML version would always create the proper CSV when the HTML version
just passes through the original form params set by the original
page). So, now when selecting all subjects automatically, we set those
values into the form less invasively (so as to not delete the dummy
value) so that we get *all* the data back instead of all the data
except for the first listed subject (which happens to be ACCT at the
moment).