Improve automatic form parsing of `<select/>`.
`<select/>` elements that do not have the multiple attribute enabled
which also have no `<option/>` marked as selected will, by the force
of the brower, automatically select the first `<option/>`. As this is
closer to browser-like behavior and some servers break when not seeing
expected parameters, this change was necessary.