Have the crawler support loading resources from geturi_facade directory.
The support is very basic. For a given URI, the last path component is
extracted and the geturi_facade directory is checked to see if it
exists and, if it does, this is loaded instead of making a real HTTP
request. The point is that some servers respond very slowly to
particular pages, making it harder to develop and test crawlers. Being
able to load such a resource in seconds locally rather than waiting
minutes for a server helps a ton.