Oracle/Ellucian crawler: add basic framework to allow generalizing the crawler.
Apparently this same Oracle/Ellucian package can produce different
columns and put interesting data in different places even for the same
report. For example, one website puts department/course/section
information in a heading rather than as a column value. This change
moves indexing into a report column out of the main crawler code so
that the exact method of extracting a particular piece of data can, in
the future, be overridden to pull the data from even a non-columnar
location.