Slightly better small screen/mobile support.
Use initial-scale=1, switch logo to SVG for infinite scaling,
remove a lot of the fixed width stuff.

Because of how I am letting the table row on input and calendar
view overflow the screen, for now I got rid of the drop shadow
that was displayed around the main area of the page. This
would display inside of the browser window and intersect with
the parts of the page that overflowed the viewport, causing
it to look weird. Hopefully in the future the input and calendar
view can support looking good on smaller screens. For now, I think
letting them overflow the viewport and letting the user pinch
zoom out yields a much better experience than we had previously.