Add CURLOPT_TCP_KEEPALIVE to avoid getting stuck.
Without this option, the crawler running on hurin.ohnopub.net
sometimes gets stuck and sits for months on end waiting for a response
from Calvin when in reality the connection is long dead. With
TCP-level keep-alive packets, such dead connections will be detected
and cause an error instead of causing the crawler to get months out of
date.