semester: Fix start/end date estimation based on semester name to use year correctly.
It was trying to use the year as the day of month and setting the year
to the current year. So we would do something like 2019-03-2007 which
would be normalized by calculating that out to something like
2025-04-05 (I do not know the exact dates and did not run that one
through the mkdate() function).

When a school is relying on this inexact date estimation, it messes up
automatic semester selection. It should be better after this change.