calvin: crawl: Fix TBA in section meetings from details page.
Some section will have a few meeting times listed and then
Slate Permutate needs to follow the link to the details page to
find other sections. The check for TBA happened prior to fetching
that details page, leading to a crash later if one of the other meetings
had a TBA. Now the check is done both before and after (to avoid fetching
the details page needlessly if TBA is known), preventing this crash.

This fixes crawling which was broken for about a month when the data
for 2020–2021 came out.

Thanks Adam DenHaan for reporting.